+++ logya/main.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# PYTHON_ARGCOMPLETE_OK
 import argparse
+import argcomplete
+
 from logya import __version__
 from logya.create import Create
 from logya.generate import Generate
@@ -54,7 +57,8 @@ def main():
     p_serve.add_argument('--port', type=int, help='server port to listen')
     p_serve.add_argument('--host', help='server host name or IP')
 
-    # process arguments
+    # activate command completion and process arguments
+    argcomplete.autocomplete(parser)
     args = parser.parse_args()
     if getattr(args, 'func', None):
         args.func(args)