+++ docs/changes.rst
@@ -6,6 +6,7 @@ Changes
 4.6.0 to 4.7.0
 --------------
 
+* Enable Tab key command completion using argcomplete.
 * Call strip on given path first in slugify.
 
 4.5.0 to 4.6.0